+++ ext_localconf.php
@@ -12,3 +12,11 @@ $GLOBALS['TYPO3_CONF_VARS']['FE']['contentRenderingTemplates'][] = 'themes/Confi
  * Add RTE configuration for theme_bootstrap4
  */
 $GLOBALS['TYPO3_CONF_VARS']['RTE']['Presets']['default'] = 'EXT:theme_bootstrap4/Configuration/RteCkeditor/Default.yaml';
+
+/*
+ * Register for hook to show preview of tt_content element of CType="themebootstrap4_xxx" in page module
+ */
+$GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['cms/layout/class.tx_cms_layout.php']['tt_content_drawItem']['themebootstrap4_text_image'] =
+    \KayStrobach\ThemeBootstrap4\Hooks\PageLayoutView\TextImagePreviewRenderer::class;
+$GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['cms/layout/class.tx_cms_layout.php']['tt_content_drawItem']['themebootstrap4_image'] =
+    \KayStrobach\ThemeBootstrap4\Hooks\PageLayoutView\ImagePreviewRenderer::class;